Add basic JavaFX GUI
