+++ .gitignore
@@ -25,3 +25,6 @@ text-ui-test/EXPECTED-UNIX.TXT
 # data files
 src/main/data/
 text-ui-test/data/
+
+# Log files
+TODO.md
