+++ README.md
@@ -39,7 +39,7 @@ If you are unsure of the path, type `cd` followed by a space and drag the folder
 If the installation is successful, you should see the [expected output](#expected-output).
 
 ## Features
-* `list`: Lists all tasks in the todo list.
+* `list`: Lists all tasks in the task list.
 * `todo <description>`: Adds a todo task with the given description.
 * `deadline <description> /by <date/time>`: Adds a deadline task with the given description and date.
 * `event <description> /from <date/time> /to <date/time>`: Adds an event task with the given description and date.
@@ -52,7 +52,7 @@ All date/time arguments must be in the format `<day>-<month>-<year> <hour><minut
 
 # Expected Output
 ```
-Loading todo list...
+Loading task list...
 Saved data not found. Creating new file to store data...
 
 ---------- Bibo says: ----------
