+++ config/checkstyle/checkstyle.xml
@@ -346,6 +346,12 @@
       <property name="throwsIndent" value="8" />
     </module>
 
+    <!-- Suppresses indentation check for lambda-style switch statements -->
+    <module name="SuppressionXpathSingleFilter">
+      <property name="checks" value="Indentation"/>
+      <property name="query" value="//SWITCH_RULE/descendant-or-self::node()"/>
+    </module>
+
     <module name="NoWhitespaceBefore">
       <!-- Checks that there is no whitespace before various unary operators. Linebreaks are allowed. -->
       <property name="tokens" value="SEMI, DOT, POST_DEC, POST_INC"/>
