+++ src/main/java/bibo/Launcher.java
@@ -1,0 +1,17 @@
+package bibo;
+
+
+import javafx.application.Application;
+
+/**
+ * A launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    /**
+     * Main method to launch the application.
+     * @param args
+     */
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
