+++ src/main/java/bibo/task/Todo.java
@@ -1,0 +1,15 @@
+package bibo.task;
+
+/**
+ * Represents a todo task.
+ */
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
