--- src/main/java/iuhiah/bibo/task/Todo.java
@@ -1,20 +1,0 @@
-package iuhiah.bibo.task;
-
-/**
- * Represents a todo task.
- */
-public class Todo extends Task {
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toFileString() {
-        return "todo" + super.toFileString();
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
