--- text-ui-test/EXPECTED.TXT
@@ -1,70 +1,0 @@
-Loading todo list...
-Successfully loaded 0 of 0 tasks.
-
----------- Bibo says: ----------
-Hello! I'm Bibo. What can I do for you today?
-
------------ You say: -----------
-list
-
----------- Bibo says: ----------
-Here is your todo list:
-
------------ You say: -----------
-todo borrow book
-
----------- Bibo says: ----------
-Got it. I've added this task:
-[T][ ] borrow book
-Now you have 1 tasks in the list.
-
------------ You say: -----------
-deadline return book /by 12-10-2021
-
----------- Bibo says: ----------
-BiboTaskDescriptionException: Invalid date/time format!
-
------------ You say: -----------
-deadline return book /by 12-12-2021 0100
-
----------- Bibo says: ----------
-Got it. I've added this task:
-[D][ ] return book (by: 12 Dec 2021 01:00 am)
-Now you have 2 tasks in the list.
-
------------ You say: -----------
-event project meeting /from 08-10-2021 1400 /to 08-10-2021 1600
-
----------- Bibo says: ----------
-Got it. I've added this task:
-[E][ ] project meeting (from: 08 Oct 2021 02:00 pm to: 08 Oct 2021 04:00 pm)
-Now you have 3 tasks in the list.
-
------------ You say: -----------
-mark 2
-
----------- Bibo says: ----------
-Alright! I've marked this task as done:
-[D][X] return book (by: 12 Dec 2021 01:00 am)
-
------------ You say: -----------
-delete 1
-
----------- Bibo says: ----------
-Alright! I've deleted this task:
-[T][ ] borrow book
-Now you have 2 tasks in the list.
-
------------ You say: -----------
-list
-
----------- Bibo says: ----------
-Here is your todo list:
-1. [D][X] return book (by: 12 Dec 2021 01:00 am)
-2. [E][ ] project meeting (from: 08 Oct 2021 02:00 pm to: 08 Oct 2021 04:00 pm)
-
------------ You say: -----------
-bye
-
----------- Bibo says: ----------
-Bye. Hope to see you again soon!
